feat: :sparkles: Got the scoreboard for `quiz` working
