+++ package.json
@@ -1,98 +1,98 @@
 {
-  "name": "compagnon",
-  "version": "2.0.0",
-  "description": "Discord bot",
-  "main": "src/index.ts",
-  "dependencies": {
-    "@discordjs/builders": "^0.5.0",
-    "@discordjs/opus": "^0.3.3",
-    "@discordjs/rest": "^0.1.0-canary.0",
-    "@joehoel/discord-reply": "^1.0.3",
-    "axios": "^0.21.1",
-    "bad-words": "^3.0.4",
-    "consola": "^2.15.0",
-    "cron": "^1.8.2",
-    "cross-fetch": "^3.0.6",
-    "discord-api-types": "^0.22.0",
-    "discord-paginationembed": "^2.1.0",
-    "discord-reply": "^0.1.2",
-    "discord-xp": "^1.1.14",
-    "discord.js": "^13.0.1",
-    "distube": "^2.8.8",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "ffmpeg-static": "^4.2.7",
-    "jest": "^26.6.3",
-    "module-alias": "^2.2.2",
-    "mongoose": "^5.11.12",
-    "ms": "^2.1.3",
-    "mysql": "^2.14.1",
-    "node-fetch": "^2.6.1",
-    "pg": "^8.5.1",
-    "redis": "^3.0.2",
-    "reflect-metadata": "^0.1.10",
-    "toml": "^3.0.0",
-    "typeorm": "0.2.29"
-  },
-  "devDependencies": {
-    "@joehoel/eslint-config": "1.0.1",
-    "@types/bad-words": "^3.0.0",
-    "@types/colors": "^1.2.1",
-    "@types/cron": "^1.7.3",
-    "@types/glob": "^7.1.3",
-    "@types/jest": "^26.0.19",
-    "@types/ms": "^0.7.31",
-    "@types/node": "16.3.1",
-    "@types/node-fetch": "^2.5.7",
-    "@types/redis": "^2.8.28",
-    "@typescript-eslint/eslint-plugin": "4.28.3",
-    "@typescript-eslint/parser": "4.28.3",
-    "cross-env": "^7.0.3",
-    "eslint": "7.30.0",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-import": "2.23.4",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-prettier": "3.4.0",
-    "eslint-plugin-react": "7.24.0",
-    "eslint-plugin-react-hooks": "4.2.0",
-    "gen-env-types": "^1.3.0",
-    "husky": "^6.0.0",
-    "jest-fetch-mock": "^3.0.3",
-    "lint-staged": "11.0.0",
-    "nodemon": "^2.0.6",
-    "pm2": "^4.5.0",
-    "prettier": "2.3.2",
-    "ts-jest": "^26.4.4",
-    "ts-node": "^9.1.1",
-    "ts-node-dev": "^1.1.1",
-    "typescript": "4.3.5"
-  },
-  "scripts": {
-    "dev": "nodemon --files src/index.ts",
-    "build": "rm -rf ./dist && tsc",
-    "start": "cross-env NODE_ENV=production node dist/index.js",
-    "test": "jest --watch",
-    "coverage": "jest --coverage",
-    "typeorm": "ts-node ./node_modules/typeorm/cli.js",
-    "lint": "eslint --fix \"./src/**/*.ts\"",
-    "prepare": "husky install",
-    "gen-env-types": "gen-env-types .env -o src/typings/env.d.ts -e ."
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "npm lint-staged"
-    }
-  },
-  "lint-staged": {
-    "src/**/*.ts": [
-      "eslint --fix",
-      "git add"
-    ]
-  },
-  "_moduleAliases": {
-    "@": "dist"
-  },
-  "keywords": [],
-  "author": "Joël Kuijper",
-  "license": "ISC"
+    "name": "compagnon",
+    "version": "2.0.0",
+    "description": "Discord bot",
+    "main": "src/index.ts",
+    "dependencies": {
+        "@discordjs/builders": "^0.5.0",
+        "@discordjs/opus": "^0.3.3",
+        "@discordjs/rest": "^0.1.0-canary.0",
+        "@joehoel/discord-reply": "^1.0.3",
+        "axios": "^0.21.1",
+        "bad-words": "^3.0.4",
+        "consola": "^2.15.0",
+        "cron": "^1.8.2",
+        "cross-fetch": "^3.0.6",
+        "discord-api-types": "^0.22.0",
+        "discord-paginationembed": "^2.1.0",
+        "discord-reply": "^0.1.2",
+        "discord-xp": "^1.1.14",
+        "discord.js": "^13.0.1",
+        "distube": "^2.8.8",
+        "dotenv": "^8.2.0",
+        "express": "^4.17.1",
+        "ffmpeg-static": "^4.2.7",
+        "jest": "^26.6.3",
+        "module-alias": "^2.2.2",
+        "mongoose": "^5.11.12",
+        "ms": "^2.1.3",
+        "mysql": "^2.14.1",
+        "node-fetch": "^2.6.1",
+        "pg": "^8.5.1",
+        "redis": "^3.0.2",
+        "reflect-metadata": "^0.1.10",
+        "toml": "^3.0.0",
+        "typeorm": "0.2.29"
+    },
+    "devDependencies": {
+        "@joehoel/eslint-config": "1.0.1",
+        "@types/bad-words": "^3.0.0",
+        "@types/colors": "^1.2.1",
+        "@types/cron": "^1.7.3",
+        "@types/glob": "^7.1.3",
+        "@types/jest": "^26.0.19",
+        "@types/ms": "^0.7.31",
+        "@types/node": "16.3.1",
+        "@types/node-fetch": "^2.5.7",
+        "@types/redis": "^2.8.28",
+        "@typescript-eslint/eslint-plugin": "4.28.3",
+        "@typescript-eslint/parser": "4.28.3",
+        "cross-env": "^7.0.3",
+        "eslint": "7.30.0",
+        "eslint-config-prettier": "8.3.0",
+        "eslint-plugin-import": "2.23.4",
+        "eslint-plugin-jsx-a11y": "6.4.1",
+        "eslint-plugin-prettier": "3.4.0",
+        "eslint-plugin-react": "7.24.0",
+        "eslint-plugin-react-hooks": "4.2.0",
+        "gen-env-types": "^1.3.0",
+        "husky": "^6.0.0",
+        "jest-fetch-mock": "^3.0.3",
+        "lint-staged": "11.0.0",
+        "nodemon": "^2.0.6",
+        "pm2": "^4.5.0",
+        "prettier": "2.3.2",
+        "ts-jest": "^26.4.4",
+        "ts-node": "^9.1.1",
+        "ts-node-dev": "^1.1.1",
+        "typescript": "4.3.5"
+    },
+    "scripts": {
+        "dev": "nodemon --files src/index.ts",
+        "build": "rm -rf ./dist && tsc",
+        "start": "cross-env NODE_ENV=production node dist/index.js",
+        "test": "jest --watch",
+        "coverage": "jest --coverage",
+        "typeorm": "ts-node ./node_modules/typeorm/cli.js",
+        "lint": "eslint --fix \"./src/**/*.ts\"",
+        "prepare": "husky install",
+        "gen-env-types": "gen-env-types .env -o src/typings/env.d.ts -e ."
+    },
+    "husky": {
+        "hooks": {
+            "pre-commit": "npm lint-staged"
+        }
+    },
+    "lint-staged": {
+        "src/**/*.ts": [
+            "eslint --fix",
+            "git add"
+        ]
+    },
+    "_moduleAliases": {
+        "@": "dist"
+    },
+    "keywords": [],
+    "author": "Joël Kuijper",
+    "license": "ISC"
 }
