+++ src/features/index.ts
@@ -7,3 +7,5 @@ export { default as dad } from "./dad";
 export { default as quiz } from "./quiz";
 export { default as interaction } from "./interaction";
 export { default as answers } from "./answers";
+export { default as checkAnswers } from "./check-answers";
+export { default as reactionrole } from "./reactionrole";
